requirements.txt
easyocr>=1.7.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
onnxruntime==1.23.2
numpy>=1.24.0
rembg>=2.0.50
openai>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
python-dotenv>=1.0.0
pydantic>=2.0.0
